compute_ctl: Refactor, moving spec_apply functions to spec_apply.rs
Seems nice to have the function and all its subroutines in the same
source file.
